wenet_asr_client/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
c-ares=1.28.1=hd590300_0
ca-certificates=2024.3.11=h06a4308_0
grpcio=1.62.1=py39h174d805_0
grpcio-tools=1.62.1=py39h3d6467e_0
ld_impl_linux-64=2.38=h1181459_1
libabseil=20240116.1=cxx17_h59595ed_2
libffi=3.4.4=h6a678d5_0
libgcc-ng=13.2.0=h807b86a_5
libgomp=13.2.0=h807b86a_5
libgrpc=1.62.1=h15f2491_0
libprotobuf=4.25.3=h08a7969_0
libre2-11=2023.09.01=h5a48ba9_2
libstdcxx-ng=13.2.0=h7e041cc_5
libzlib=1.2.13=hd590300_5
ncurses=6.4=h6a678d5_0
openssl=3.2.1=hd590300_1
pip=23.3.1=py39h06a4308_0
portaudio=19.6.0=h7b6447c_4
protobuf=4.25.3=py39h1be52a0_0
pyaudio=0.2.11=py39h27cfd23_2
python=3.9.18=h955ad1f_0
python_abi=3.9=2_cp39
re2=2023.09.01=h7f4b329_2
readline=8.2=h5eee18b_0
setuptools=68.2.2=py39h06a4308_0
sqlite=3.41.2=h5eee18b_0
tk=8.6.12=h1ccaba5_0
tzdata=2024a=h04d1e81_0
wheel=0.41.2=py39h06a4308_0
xz=5.4.6=h5eee18b_0
zlib=1.2.13=hd590300_5
